Comment (no changes)
From what I read so far, the only way to modify an existing txt file is to replace it or completely write over the existing file. As a result, I think that instead of one giant file with all student information (what we seem to have so far), a bunch of individual student files might be better so we don't read/write the entire student body every time we change something. I'm not sure how to properly name the files, though...
"data" + File.separator + "students" + File.separator + currentStudent.getUserName() + ".txt" ? Probably should be a student ID which is guaranteed to be unique (add that under the "nice to haves"). Also, initializing the model object will have to be changed so that it will load the student's unique txt file after login.
(There's so much going on I'm afraid to change anything myself!)